Damage System for Towers

+++ Assets/_Game/Scripts/DamageSystem/IDamageable.cs
@@ -2,6 +2,6 @@
 {
     public interface IDamageable
     {
-        void Hurt(int amount);
+        void Hurt(float amount);
     }
 }
